Modify study goal page and related page code
