+++ Growing-Traveler/Growing-Traveler/Model/Enum/StatusType.swift
@@ -1,0 +1,32 @@
+//
+//  StatusType.swift
+//  Growing-Traveler
+//
+//  Created by Jenny Hung on 2022/5/12.
+//
+
+import UIKit
+
+enum StatusType {
+    
+    case pending
+    
+    case running
+    
+    case finished
+    
+    var title: String {
+        
+        switch self {
+            
+        case .pending: return "待處理"
+            
+        case .running: return "處理中"
+            
+        case .finished: return "已處理"
+            
+        }
+        
+    }
+    
+}
